remove old landing page routes
