+++ api/api/services/features.py
@@ -7,13 +7,13 @@ from typing import Any, NamedTuple
 from pydantic import BaseModel
 
 from api.services.internal_tasks._internal_tasks_utils import OFFICIALLY_SUGGESTED_TOOLS, officially_suggested_tools
-from api.services.internal_tasks.agent_suggestions_service import (
-    get_supported_task_input_types,
-    get_supported_task_output_types,
-)
 from api.services.slack_notifications import SlackNotificationDestination, get_user_and_org_str, send_slack_notification
 from core.agents.agent_output_example import SuggestedAgentOutputExampleInput, stream_suggested_agent_output_example
 from core.agents.agent_suggestion_validator_agent import SuggestedAgentValidationInput, run_suggested_agent_validation
+from core.agents.chat_task_schema_generation.chat_task_schema_generation_task import (
+    InputSchemaFieldType,
+    OutputSchemaFieldType,
+)
 from core.agents.chat_task_schema_generation.chat_task_schema_generation_task_utils import build_json_schema_with_defs
 from core.agents.chat_task_schema_generation.schema_generation_agent import (
     SchemaBuilderInput,
@@ -42,6 +42,15 @@ from core.tools.browser_text.browser_text_tool import fetch_url_content_scraping
 from core.tools.search.run_perplexity_search import stream_perplexity_search
 from core.utils.iter_utils import safe_map
 from core.utils.schema_utils import json_schema_from_json
+
+
+def get_supported_task_input_types() -> list[str]:
+    return [type.value for type in InputSchemaFieldType] + ["enum", "array", "object"]
+
+
+def get_supported_task_output_types() -> list[str]:
+    return [type.value for type in OutputSchemaFieldType] + ["enum", "array", "object"]
+
 
 _logger = logging.getLogger(__name__)
